+++ apps/gym/src/components/reusable/Dropdown/Dropdown.tsx
@@ -1,15 +1,10 @@
 'use client'
 
 import styles from './Dropdown.module.css'
+import { DropdownProps } from '@/types'
 import { Button } from './Button/Button'
 import { Content } from './Content/Content'
 import { Item } from './Item/Item'
-import { CSSProperties, HTMLAttributes, ReactNode } from 'react'
-
-export interface DropdownProps extends HTMLAttributes<HTMLDivElement> {
-  sx?: CSSProperties
-  children: ReactNode | ReactNode[]
-}
 
 export const Dropdown = ({ sx, children }: DropdownProps) => {
   return (